Ammo tick polish
Ammo ticks will use bar if the ticks would exceed a certain threshold of the weapon panel's space (even if the data does not indicate use_bar). This is based on the size of the ammo tick texture.
Weapons that use an ammo bar instead of ticks (including through the overflow method described above) will show the mag counter in addition to the reserves counter
Fixed "Real weapon ammo" giving incorrect ammo readout in HUD